pyproject.toml
[project]
name = "invoice-scraper"
version = "0.1.0"
description = "Automated invoice download system using Playwright"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.40.0",
    "pytest-playwright>=0.4.0",
    "pytest-asyncio>=0.24.0",
    "google-cloud-secret-manager>=2.0.0",
    "google-auth>=2.0.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.9.0",
    "pytest>=9.0.2",
    "pyotp>=2.9.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 140
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "PL", "G", "TRY"]
ignore = [
    "D100",     # missing-module-docstring
    "D101",     # missing-class-docstring
    "D103",     # missing-function-docstring
    "G004",     # logging-fstring-interpolation
    "TRY002",   # broad-exception-raised (raise-vanilla-class)
    "TRY003",   # long exception messages (commonly triggered with TRY)
    "TRY300",   # try-consider-else
    "PLR0913",  # too-many-arguments
    "PLR2004",  # magic-value-comparison
    "PLR0914",  # too-many-locals
    "PLR0917",  # too-many-positional-arguments
]
